# userservice/dto/src/main/resources/openapi/user.yaml
openapi: 3.0.3
info:
  title: User Service API
  version: 1.0.0
  description: API for managing users
paths:
  /api/v1/users:
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequestDTO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
    get:
      summary: Get all users
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDTO'
  /api/v1/users/{id}:
    put:
      summary: Update user
      operationId: updateUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequestDTO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDTO'
    delete:
      summary: Delete user
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
components:
  schemas:
    UserRole:
      type: string
      enum: [USER, ADMIN]
    UserCreateRequestDTO:
      type: object
      required: [email, password, name, surname, phone, role]
      properties:
        email:
          type: string
          format: email
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 255
        name:
          type: string
          maxLength: 30
        surname:
          type: string
          maxLength: 30
        phone:
          type: string
          maxLength: 15
        role:
          $ref: '#/components/schemas/UserRole'
    UserUpdateRequestDTO:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 100
        name:
          type: string
          maxLength: 30
        surname:
          type: string
          maxLength: 30
        phone:
          type: string
          maxLength: 15
        role:
          $ref: '#/components/schemas/UserRole'
    UserResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        surname:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time